requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
aiofiles==24.1.0
python-multipart==0.0.20
python-dotenv
python-dotenv
